gcp_functions/http_requirements.txt
functions-framework==3.*
google-api-core==1.31.1
google-cloud-pubsub==2.7.0